break out StatusImage for reuse
