+++ app/javascript/mastodon/features/compose/redesign/attachments.module.scss
@@ -34,11 +34,6 @@
   overflow: hidden;
 }
 
-.blurHash {
-  width: 100%;
-  height: 100%;
-}
-
 .mediaUploadPending {
   animation: pulse 1s ease-in-out infinite;
   background-color: var(--color-bg-disabled);
